fix unreachable objects being created when new swapchain created without fully deleting old swapchain

+++ core/renderer/renderer.cpp
@@ -46,7 +46,6 @@ void Renderer::framebufferResizeCallback(GLFWwindow* window, int width, int heig
 
 void Renderer::initVulkan() {
 	createInstanceAndDevice();
-	createSurface();
 	createSwapChain();
 	createImageViews();
 	createRenderPass();
@@ -148,7 +147,7 @@ void Renderer::updateUniformBuffer(uint32_t current) {
 
 }
 
-void Renderer::cleanupSwapChain() {
+void Renderer::cleanupSwapChain(bool destroySwapchain) {
 	for (auto framebuffer : swapChainFramebuffers) {
 		vkDestroyFramebuffer(logicalDevice, framebuffer, nullptr);
 	}
@@ -156,26 +155,24 @@ void Renderer::cleanupSwapChain() {
 	for (auto imageView : swapChainImageViews) {
 		vkDestroyImageView(logicalDevice, imageView, nullptr);
 	}
-
-	vkb::destroy_swapchain(swapChain);
+	if (destroySwapchain) vkb::destroy_swapchain(swapChain);
 }
 
 void Renderer::Cleanup() {
 	cleanupSwapChain();
 
-	vkDestroyDescriptorPool(logicalDevice, descriptorPool, nullptr);
-	vkDestroyDescriptorSetLayout(logicalDevice, descriptorSetLayout, nullptr);
-
 	vkDestroyPipeline(logicalDevice, graphicsPipeline, nullptr);
 	vkDestroyPipelineLayout(logicalDevice, pipelineLayout, nullptr);
-
 	vkDestroyRenderPass(logicalDevice, renderPass, nullptr);
 
-	for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++) {
-		vkDestroySemaphore(logicalDevice, renderFinishedSemaphores[i], nullptr);
-		vkDestroySemaphore(logicalDevice, imageAvailableSemaphores[i], nullptr);
-		vkDestroyFence(logicalDevice, inFlightFences[i], nullptr);
+	for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
+	{
+		vkDestroyBuffer(logicalDevice, uniformBuffers[i], nullptr);
+		vkFreeMemory(logicalDevice, uniformBuffersMemory[i], nullptr);
 	}
+
+	vkDestroyDescriptorPool(logicalDevice, descriptorPool, nullptr);
+	vkDestroyDescriptorSetLayout(logicalDevice, descriptorSetLayout, nullptr);
 
 	vkDestroyBuffer(logicalDevice, indexBuffer, nullptr);
 	vkFreeMemory(logicalDevice, indexBufferMemory, nullptr);
@@ -183,12 +180,16 @@ void Renderer::Cleanup() {
 	vkDestroyBuffer(logicalDevice, vertexBuffer, nullptr);
 	vkFreeMemory(logicalDevice, vertexBufferMemory, nullptr);
 
+	resources::Shader* vertShader = ResourceLoader::Load<resources::Shader>("shaders/shader.vert");
+	resources::Shader* fragShader = ResourceLoader::Load<resources::Shader>("shaders/shader.frag");
+	vkDestroyShaderModule(logicalDevice, vertShader->GetShaderModule(logicalDevice), nullptr);
+	vkDestroyShaderModule(logicalDevice, fragShader->GetShaderModule(logicalDevice), nullptr);
 
 	for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++) {
-		vkDestroyBuffer(logicalDevice, uniformBuffers[i], nullptr);
-		vkFreeMemory(logicalDevice, uniformBuffersMemory[i], nullptr);
+		vkDestroySemaphore(logicalDevice, renderFinishedSemaphores[i], nullptr);
+		vkDestroySemaphore(logicalDevice, imageAvailableSemaphores[i], nullptr);
+		vkDestroyFence(logicalDevice, inFlightFences[i], nullptr);
 	}
-
 
 	vkDestroyCommandPool(logicalDevice, commandPool, nullptr);
 
@@ -197,20 +198,13 @@ void Renderer::Cleanup() {
 		DestroyDebugUtilsMessengerEXT(_instance, debugMessenger, nullptr);
 	}
 
-	vkb::destroy_surface(_instance, surface);
+	vkDestroySurfaceKHR(_instance, surface, nullptr);
 	vkb::destroy_instance(_instance);
 
 	glfwDestroyWindow(window);
 
 	glfwTerminate();
 }
-
-void Renderer::createSurface() {
-	if (glfwCreateWindowSurface(_instance, window, nullptr, &surface) != VK_SUCCESS) {
-		Log.FatalError("Vulkan", "Failed to create window surface.");
-	}
-}
-
 
 void Renderer::createSwapChain() {
 	unsigned int imageCount = 2;
@@ -246,6 +240,7 @@ void Renderer::recreateSwapChain() {
 	}
 
 	vkDeviceWaitIdle(logicalDevice);
+	cleanupSwapChain(false);
 	createSwapChain();
 	createImageViews();
 	createFramebuffers();
